CAG_DocuAI/requirements.txt
Flask==2.3.3
Flask_Cors==4.0.0
numpy==1.25.2
oci==2.110.2
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.0.0
Pillow==10.0.0
python-dotenv==1.0.0
Requests==2.31.0
